+++ src/cjs/PlotBoilerplate.js
@@ -190,6 +190,7 @@ var PlotBoilerplate = /** @class */ (function () {
      * @param {boolean=} [config.enableGL=false] - Indicates if the application should use the experimental WebGL features (not recommended).
      * @param {boolean=} [config.enableSVGExport=true] - Indicates if the SVG export should be enabled (default is true).
      *                                                   Note that changes from the postDraw hook might not be visible in the export.
+     * @param {string=} [config.title=null] - Specify any hover tile here. It will be attached as a `title` attribute to the most elevated element.
      */
     function PlotBoilerplate(config) {
         /**
@@ -404,6 +405,11 @@ var PlotBoilerplate = /** @class */ (function () {
         }
         else {
             throw "Element is neither a canvas nor an svg element.";
+        }
+        // At this point the event cacher element is deinfed and located at highest elevation.
+        // Set `title` attribut?
+        if (config.title) {
+            this.eventCatcher.setAttribute("title", config.title);
         }
         this.draw.scale.set(this.config.scaleX, this.config.scaleY);
         this.fill.scale.set(this.config.scaleX, this.config.scaleY);