+++ src/ts/interfaces/core.ts
@@ -14,6 +14,7 @@
  * @modified 2021-05-31 Splitted the large interfaces.ts file into this one and others.
  * @modified 2021-06-21 Added `IBounds.getCenter()`.
  * @modified 2021-11-16 Added `text` options to the `DrawConfig`.
+ * @modified 2022-08-01 Added `title` param to the `PBParams` interface.
  **/
 
 import { Vertex } from "../Vertex";
@@ -156,8 +157,8 @@ export interface Config {
  * For initialization the constructor needs a mix of config and draw-settings.
  */
 export interface PBParams extends Config {
-  // , DrawConfig {
-  // No additional attributes, just merge two interfaces.
+  title?: string; // Specify any hover tile here. It will be attached as a `title` attribute to the most elevated element.
+  // No more additional attributes, just merge two interfaces.
 }
 
 export interface DrawSettings {